Move querying view permissions into a conditional component
