+++ src/components/AppCard/index.tsx
@@ -1,6 +1,6 @@
 import { type FC } from 'react'
 import { Link } from 'react-router-dom'
-import { useGetRuntimeRoflmarketInstances, type RoflApp } from '../../nexus/api'
+import { type RoflApp } from '../../nexus/api'
 import { Badge } from '@oasisprotocol/ui-library/src/components/ui/badge'
 import { Button } from '@oasisprotocol/ui-library/src/components/ui/button'
 import { Card, CardContent, CardFooter, CardHeader } from '@oasisprotocol/ui-library/src/components/ui/card'
@@ -10,8 +10,7 @@ import { cn } from '@oasisprotocol/ui-library/src/lib/utils'
 import { trimLongString } from '../../utils/trimLongString'
 import { formatDistanceToNow, parseISO } from 'date-fns'
 import { useAccount } from 'wagmi'
-import { hasViewLogsPermission } from '../../utils/hasViewLogsPermission'
-import { isMachineRemoved } from '../MachineStatusIcon/isMachineRemoved'
+import { ViewWithOnlyLogsPermission } from './ViewWithOnlyLogsPermission'
 
 type AppCardProps = {
   app: RoflApp
@@ -21,17 +20,6 @@ type AppCardProps = {
 
 export const AppCard: FC<AppCardProps> = ({ app, network, type }) => {
   const { address } = useAccount()
-  const machinesQuery = useGetRuntimeRoflmarketInstances(network, 'sapphire', {
-    limit: 100,
-    offset: 0,
-    deployed_app_id: app.id,
-  })
-  const machinesWithOnlyLogsPermission =
-    type === 'explore' && address && machinesQuery.data?.data.instances
-      ? machinesQuery.data.data.instances
-          .filter(machine => !isMachineRemoved(machine))
-          .filter(machine => hasViewLogsPermission(machine, address))
-      : []
 
   return (
     <Card className="rounded-md">
@@ -84,22 +72,9 @@ export const AppCard: FC<AppCardProps> = ({ app, network, type }) => {
           </Button>
         )}
 
-        {machinesWithOnlyLogsPermission.length > 0 && (
-          <div className="flex flex-col gap-2">
-            {machinesWithOnlyLogsPermission.map(machine => (
-              <Button key={machine.id} variant="default" asChild>
-                {/* TODO: Link to logs subpage. But shadcn does not support routed tabs nicely. */}
-                <Link to={`/dashboard/machines/${machine.provider}/instances/${machine.id}`}>View logs</Link>
-              </Button>
-            ))}
-          </div>
-        )}
+        {type === 'explore' && address && <ViewWithOnlyLogsPermission app={app} network={network} />}
 
-        <Button
-          variant="secondary"
-          asChild
-          className="bg-background"
-        >
+        <Button variant="secondary" asChild className="bg-background">
           <a
             href={`https://explorer.oasis.io/${network}/sapphire/rofl/app/${app.id}`}
             target="_blank"
